+++ packages/chlu/rollup.config.js
@@ -1,31 +1,9 @@
 import builtins from 'rollup-plugin-node-builtins'
-import resolve from 'rollup-plugin-node-resolve'
-import commonjs from 'rollup-plugin-commonjs'
-import uglify from 'rollup-plugin-uglify'
 import babel from 'rollup-plugin-babel'
 import json from 'rollup-plugin-json'
-import { minify } from 'uglify-es'
 import pkg from './package.json'
 
 export default [
-  // browser-friendly UMD build
-  {
-    input: 'src/main.js',
-    output: {
-      file: pkg.browser,
-      format: 'umd',
-      name: 'chlu',
-    },
-    plugins: [
-      builtins(),
-      json(),
-      resolve(), // so Rollup can find deps
-      commonjs(), // so Rollup can convert deps to ES modules
-      babel(),
-      uglify({}, minify),
-    ],
-  },
-
   // Builds: CommonJS (for Node) and ES module (for bundlers)
   {
     input: 'src/main.js',
@@ -45,7 +23,7 @@ export default [
       'lodash.min',
       'lodash.reverse',
       'lodash.trim',
-      'posthtml-ast-contains-only-empty-space',
+      'ast-contains-only-empty-space',
       'semver-compare',
       'split-lines',
     ],